Put stuff back where it was again
